requirements.txt
streamlit
openai
python-dotenv
PyPDF2
spacy
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
